requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
sqlalchemy==1.4.22
psycopg2-binary==2.9.1
mysql-connector-python==8.0.26
pandas==1.3.3
scikit-learn==1.0
sentence-transformers==2.1.0
PyPDF2==1.26.0
python-docx==0.8.11
python-dotenv==1.0.0
google-generativeai==0.3.1